Fix compatibility with older browser versions
Mixing `const` and `var` can trigger bugs in older browsers, e.g. Chrome 53.

To be precise when executing this code with mixed `const` and `var` in Chrome 53
as part of a webpack bundle with a webpack `devtool` setting, the code will throw
"Uncaught ReferenceError: styles is not defined" in the `getDefaultProps` function.